manager/requirements.txt
SQLAlchemy==1.1.15
alembic==0.9.6
grpcio==1.7.3